refactor: updated the SusAnalytics.vue by creating charts as separate child component
