+++ src/views/admin/SusAnalytics.vue
@@ -93,65 +93,12 @@
             <v-row>
               <!-- Histogram Chart -->
               <v-col cols="12" md="6">
-                <div class="chart-container">
-                  <h3 class="text-h6 mb-4">Score Distribution Histogram</h3>
-                  <canvas ref="histogramCanvas" style="max-height: 300px;"></canvas>
-                </div>
+                <SusHistogramChart :scores="tasksArray.map(r => r.susScore)" />
               </v-col>
-              
+
               <!-- Horizontal Bar Distribution -->
               <v-col cols="12" md="6">
-                <div class="chart-container rating-chart-container">
-                  <h3 class="text-h6 mb-4">Rating Categories</h3>
-                  <div class="distribution-chart" style="height: 300px; display: flex; flex-direction: column; justify-content: space-around;">
-                    <div 
-                      v-for="(item, index) in distributionData" 
-                      :key="index"
-                      class="distribution-row"
-                      style="margin-bottom: 20px;"
-                    >
-                      <div class="d-flex align-center">
-                        <div class="rating-label" style="width: 80px;">
-                          {{ item.label }}
-                        </div>
-                        <div class="flex-grow-1 mx-4">
-                          <div class="progress-container">
-                            <div 
-                              class="progress-bar"
-                              :style="{
-                                width: `${item.percentage}%`,
-                                backgroundColor: item.color,
-                                borderRadius: '20px',
-                                height: '32px',
-                                position: 'relative',
-                                minWidth: item.count > 0 ? '40px' : '0px'
-                              }"
-                            >
-                              <span 
-                                v-if="item.count > 0"
-                                class="progress-text"
-                                style="
-                                  position: absolute;
-                                  left: 50%;
-                                  top: 50%;
-                                  transform: translate(-50%, -50%);
-                                  color: white;
-                                  font-weight: bold;
-                                  font-size: 12px;
-                                "
-                              >
-                                {{ item.count }}
-                              </span>
-                            </div>
-                          </div>
-                        </div>
-                        <div class="score-range text-caption text-grey" style="width: 60px; text-align: right;">
-                          {{ item.range }}
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </div>
+                <SusRatingChart :scores="tasksArray.map(r => r.susScore)" />
               </v-col>
             </v-row>
           </v-card-text>
@@ -329,39 +276,18 @@
 </template>
 
 <script setup>
-import { ref, computed, onMounted, nextTick } from 'vue'
+import { ref, computed } from 'vue'
 import { getSUSRating, getRatingColor, calculateSUSScore } from '@/utils/susCalculator'
-import {
-  Chart as ChartJS,
-  CategoryScale,
-  LinearScale,
-  BarElement,
-  BarController,
-  Title,
-  Tooltip,
-  Legend,
-} from 'chart.js'
 import { useStore } from 'vuex';
+import SusHistogramChart from '@/components/atoms/SusHistogramChart.vue';
+import SusRatingChart from '@/components/atoms/SusRatingChart.vue';
 
 const store = useStore();
-
-ChartJS.register(
-  CategoryScale,
-  LinearScale,
-  BarElement,
-  BarController,
-  Title,
-  Tooltip,
-  Legend
-)
 
 const detailsModal = ref(false)
 const selectedResponse = ref(null)
 const selectedRatingFilter = ref(null)
 const scoreRange = ref([0, 100])
-
-const histogramCanvas = ref(null)
-let histogramChart = null
 
 const tableHeaders = [
   { title: 'User', key: 'user', sortable: true },
@@ -426,98 +352,6 @@ const filteredResponses = computed(() => {
   return filtered
 })
 
-const distributionData = computed(() => {
-  const scores = tasksArray.value.map(r => r.susScore)
-  const total = scores.length
-
-  const ranges = [
-    { label: 'Poor', range: '0-51', min: 0, max: 51, color: '#f44336' },
-    { label: 'OK', range: '52-67', min: 52, max: 67, color: '#ff9800' },
-    { label: 'Good', range: '68-83', min: 68, max: 83, color: '#2196f3' },
-    { label: 'Excellent', range: '84-100', min: 84, max: 100, color: '#4caf50' }
-  ]
-
-  const maxCount = Math.max(...ranges.map(range => scores.filter(score => score >= range.min && score <= range.max).length))
-
-  return ranges.map(range => {
-    const count = scores.filter(score => score >= range.min && score <= range.max).length
-    return {
-      ...range,
-      count,
-      percentage: maxCount > 0 ? (count / maxCount) * 100 : 0
-    }
-  })
-})
-
-const histogramData = computed(() => {
-  const scores = tasksArray.value.map(r => r.susScore)
-  const buckets = [
-    { label: '0-19', min: 0, max: 19, color: '#f44336' },
-    { label: '20-39', min: 20, max: 39, color: '#f44336' },
-    { label: '40-59', min: 40, max: 59, color: '#ff9800' },
-    { label: '60-79', min: 60, max: 79, color: '#2196f3' },
-    { label: '80-100', min: 80, max: 100, color: '#4caf50' }
-  ]
-
-  const bucketCounts = buckets.map(bucket => scores.filter(score => score >= bucket.min && score <= bucket.max).length)
-
-  return {
-    labels: buckets.map(b => b.label),
-    datasets: [{
-      label: 'Number of Users',
-      data: bucketCounts,
-      backgroundColor: buckets.map(b => b.color + '80'),
-      borderColor: buckets.map(b => b.color),
-      borderWidth: 1,
-      borderRadius: 4,
-      borderSkipped: false,
-    }]
-  }
-})
-
-const histogramOptions = {
-  responsive: true,
-  maintainAspectRatio: false,
-  plugins: {
-    legend: { display: false },
-    tooltip: {
-      callbacks: {
-        title: function(context) {
-          return `Score Range: ${context[0].label}`
-        },
-        label: function(context) {
-          const count = context.parsed.y
-          return `${count} user${count !== 1 ? 's' : ''}`
-        }
-      }
-    }
-  },
-  scales: {
-    x: {
-      title: {
-        display: true,
-        text: 'SUS Score Ranges',
-        font: { size: 14, weight: 'bold' }
-      },
-      grid: { display: false }
-    },
-    y: {
-      title: {
-        display: true,
-        text: 'Number of Users',
-        font: { size: 14, weight: 'bold' }
-      },
-      beginAtZero: true,
-      ticks: {
-        stepSize: 1,
-        callback: function(value) {
-          return Number.isInteger(value) ? value : ''
-        }
-      }
-    }
-  }
-}
-
 function openDetailsModal(response) {
   selectedResponse.value = response
   detailsModal.value = true
@@ -540,25 +374,6 @@ function getResponseLabel(score) {
     default: return 'Unknown'
   }
 }
-
-function createHistogramChart() {
-  if (histogramCanvas.value) {
-    if (histogramChart) histogramChart.destroy()
-    const ctx = histogramCanvas.value.getContext('2d')
-    if (ctx) {
-      histogramChart = new ChartJS(ctx, {
-        type: 'bar',
-        data: histogramData.value,
-        options: histogramOptions
-      })
-    }
-  }
-}
-
-onMounted(async () => {
-  await nextTick()
-  createHistogramChart()
-})
 </script>
 
 <style scoped>
@@ -575,45 +390,9 @@ onMounted(async () => {
   font-weight: 600;
 }
 
-.distribution-chart {
-  width: 100%;
-}
-
-.distribution-row {
-  width: 100%;
-}
-
-.progress-container {
-  background-color: #f5f5f5;
-  border-radius: 20px;
-  height: 24px;
-  position: relative;
-  overflow: hidden;
-}
-
-.progress-bar {
-  transition: width 0.3s ease;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.rating-label {
-  font-weight: 500;
-  font-size: 14px;
-}
-
 .response-item {
   background: #fafafa !important;
   border: 1px solid #e0e0e0 !important;
-}.chart-container {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.chart-container canvas {
-  flex-grow: 1;
 }
 
 .question-text {
@@ -623,27 +402,5 @@ onMounted(async () => {
 .response-score {
   min-width: 120px;
   justify-content: flex-end;
-}
-
-.chart-container {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.chart-container canvas {
-  flex-grow: 1;
-}
-
-.rating-chart-container {
-  min-height: 350px;
-}
-
-.distribution-chart {
-  padding: 20px 0;
-}
-
-.progress-container {
-  height: 32px !important;
 }
 </style>
